Not preserving block gap if the connecting edge is of type COND_NOT_TAKEN
I am seeing some edges of type COND_TAKEN are incorrectly labeled as
COND_NOT_TAKEN when compiling the cruise controller binary
(https://github.umn.edu/taejoon/component-based-assurance/tree/master/coverage/exp250-in/cruise)
with the -O2 optimization option with gcc. This causes the assertion
(target->start() >= block_->end()) to not be satisfied in
dyninstAPI/src/Relocation/CFG/RelocBlock.C:preserveBlockGap(). Since
this is primarily intended for defensive binaries, which is not
something we are using, removing assertion for edges of type
COND_NOT_TAKEN for now.

The real fix would be to figure out why dyninst labels these edges
incorrectly. For now, I am seeing the root cause of these incorrectly
labeled edges arising out of the branch (if (L104)) in cruise.c. But
there are ~100 or so branches before this branch in the top_step
function, making it difficult to make a smaller example demonstrating
this issue.